ref[page]: create template for reusability
