+++ src/components/organisms/executiveSummary/index.ts
@@ -1,3 +1,2 @@
-export { default as ExecutiveSummaryIntro } from '../introSection/IntroSection'
 export { default as NextSection } from '../nextSection/NextSection'
 export { default as Body } from './Body'
