+++ src/components/organisms/footer/index.ts
@@ -1,0 +1,1 @@
+export { default as FooterContent } from './FooterContent'
