+++ src/components/templates/index.ts
@@ -1,3 +1,4 @@
+export { default as ArticlePageTemplate } from './articlePageTemplate/ArticlePageTemplate'
 export * from './MailChimp'
 export { default as SectionTemplate } from './Section'
 export { default as TwoColumnStatTextTemplate } from './twoColumnTemplate/TwoColumnTemplate'
